refactor: split quickstart lesson into steps, fold, and surface modules
Keep scripts/quickstart.ts as a stable re-export entrypoint so docs,
llms, and consumer-compat imports do not change. Move lesson content,
foldSakura, and agent/page surface helpers into scripts/quickstart/*,
and teach CI path routing + content-hash inputs about the new tree so
module edits still schedule docs, journeys, and consumer jobs.
