+++ scripts/ci-routing/content-hash.ts
@@ -113,6 +113,7 @@ const DOCS_SURFACE_CONTENT_INPUTS: readonly string[] = [
   "scripts/cli-docs.ts",
   "scripts/guide-code-contract.ts",
   "scripts/quickstart.ts",
+  "scripts/quickstart/**",
   "scripts/highlight-code.ts",
   "scripts/highlight-code.test.ts",
   "scripts/deployment-artifact.ts",
@@ -292,6 +293,7 @@ export const JOB_CONTENT_INPUTS: Record<CacheableExecution, readonly string[]> =
     "scripts/deployment-smoke-cli.ts",
     "scripts/deployment-smoke.test.ts",
     "scripts/quickstart.ts",
+    "scripts/quickstart/**",
     "scripts/gen-manifest.ts",
     "scripts/gen-lifecycle.ts",
   ],
@@ -304,6 +306,7 @@ export const JOB_CONTENT_INPUTS: Record<CacheableExecution, readonly string[]> =
     "scripts/consumer-compat.test.ts",
     "scripts/guide-code-contract.ts",
     "scripts/quickstart.ts",
+    "scripts/quickstart/**",
     "scripts/quickstart-timing.ts",
     "scripts/quickstart-timing.test.ts",
     "scripts/support-matrix.ts",
